+++ src/pipecat/services/google/gemini_live/__init__.py
@@ -1,3 +1,3 @@
 from .file_api import GeminiFileAPI
 from .llm import GeminiLiveLLMService
-from .vertex import GeminiLiveVertexLLMService
+from .llm_vertex import GeminiLiveVertexLLMService
